Update: tara_id de float a int

+++ src/modules/lineas_entrada_salida_service/infrastructure/api/schemas/lineas_salida.py
@@ -29,4 +29,4 @@ class LineasSalidaPaginatedResponse(BaseModel):
     data: List[LineasSalidaResponse]
 
 class TaraIdRequest(BaseModel):
-    tara_id: float
+    tara_id: int
